Add add_block_container() method
