+++ plugins/woocommerce/src/Internal/Admin/BlockTemplates/BlockTemplate.php
@@ -83,11 +83,4 @@ class BlockTemplate implements BlockTemplateInterface {
 	public function &get_parent(): ?BlockContainerInterface {
 		return null;
 	}
-
-	/**
-	 * Get the template as a formatted template.
-	 */
-	public function get_formatted_template(): array {
-		return $this->get_inner_blocks_formatted_template();
-	}
 }
